Task_04/study-agent/pyproject.toml
[project]
name = "study-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.6.1",
    "pypdf>=6.4.0",
    "streamlit>=1.51.0",
    "python-dotenv>=1.0.1",
]

